+++ src/itemChoice/index.tsx
@@ -25,7 +25,7 @@ interface TypeProps {
 export interface ItemChoiceProps {
   readonly className?: Classcat.Class,
   readonly href?: string | JSX.Element,
-  readonly key?: string | number,
+  readonly id?: string | number,
   readonly label?: string,
   readonly subLabel?: string,
   readonly children?: React.ReactNode,
@@ -53,7 +53,7 @@ class ItemChoice extends PureComponent <ItemChoiceProps> {
     const {
       children, className, highlighted, selected, status,
       onClick, onBlur, onFocus, onMouseDown, href, label, subLabel, leftAddon, rightAddon,
-      onDoneAnimationEnd, key,
+      onDoneAnimationEnd, id,
     } = this.props
     const classNames = cc([prefix({
       itemChoice: true,
@@ -107,7 +107,7 @@ class ItemChoice extends PureComponent <ItemChoiceProps> {
     return (
       <Component
         role="option"
-        key={key}
+        key={id}
         aria-selected={selected}
         {...typeProps}
       >