+++ src/menu/index.unit.tsx
@@ -1,21 +1,21 @@
 import React from 'react'
 import Menu from 'menu'
+import ItemChoice from 'itemChoice'
+import { HomeIcon, NewspaperIcon } from 'icon'
 
 const defaultProps = {
   items: [
     {
       id: 'menu-item-1',
-      label: 'Menu item 1',
-      url: '#',
-    },
-    {
-      id: 'separator-1',
-      separator: true,
+      label: 'Dashboard',
+      leftAddon: <HomeIcon />,
+      href: '/',
     },
     {
       id: 'menu-item-2',
-      label: 'Menu item 2',
-      url: '#',
+      label: 'Rides offered',
+      leftAddon: <NewspaperIcon />,
+      href: '/rides',
     },
   ],
 }
@@ -29,7 +29,9 @@ describe('Menu', () => {
 
   it('Should render menu items', () => {
     const wrapper = mount(<Menu {...defaultProps} />)
-    expect(wrapper.find('li')).toHaveLength(3)
-    expect(wrapper.find('.separator')).toHaveLength(1)
+    expect(wrapper.find(ItemChoice)).toHaveLength(2)
+    expect(wrapper.find(ItemChoice).first().prop('href')).toEqual('/')
+    expect(wrapper.find(ItemChoice).first().text()).toEqual('Dashboard')
+    expect(wrapper.find(HomeIcon)).toHaveLength(1)
   })
 })